+++ R/f_likelyPlatformTrial.R
@@ -48,8 +48,6 @@
 #' # fields needed
 #' f.likelyPlatformTrial()
 #'
-#' \dontrun{
-#'
 #' # apply trial concept when creating data frame
 #' dbc <- nodbi::src_sqlite(
 #'   dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
@@ -57,8 +55,7 @@
 #' trialsDf <- dbGetFieldsIntoDf(
 #'   calculate = "f.likelyPlatformTrial",
 #'   con = dbc)
-#' }
-#'
+#' trialsDf
 #'
 f.likelyPlatformTrial <- function(df = NULL) {
 
@@ -193,7 +190,6 @@ f.likelyPlatformTrial <- function(df = NULL) {
     dplyr::filter(.data$name != "EUCTR") %>% # since this has no country suffix
     dplyr::filter(.data$name != "ctrname") %>%
     dplyr::filter(.data$value != "") %>%
-    # TODO actually broaden this column to include _id to have all identifiers together
     dplyr::filter(.data$value != .data$`_id`) %>%
     dplyr::select(!"name") %>%
     unique() %>%
@@ -283,7 +279,8 @@ f.likelyPlatformTrial <- function(df = NULL) {
     ),
     #
     # identify rows of possibly related trials
-    # TODO add 'contained in' e.g. RAINBO-p53abnRED, RAINBO
+    # 'contained in' e.g. RAINBO-p53abnRED, RAINBO
+    # is realised by using Jaro Winkler distance
     .maybeRelatedTrials = indexSimilarX(.data$titleRefs),
     #
     # turn row indices to ids of possibly related trials